+++ packages/fpdart/lib/src/date.dart
@@ -2,21 +2,6 @@ import 'io.dart';
 import 'typeclass/eq.dart';
 import 'typeclass/order.dart';
 
-/// `fpdart` extension methods on [DateTime]
-extension FpdartDateTime on DateTime {
-  /// Return `true` when this [DateTime] and `other` have the same **year**.
-  bool eqvYear(DateTime other) => dateEqYear.eqv(this, other);
-
-  /// Return `true` when this [DateTime] and `other` have the same **month**.
-  bool eqvMonth(DateTime other) => dateEqMonth.eqv(this, other);
-
-  /// Return `true` when this [DateTime] and `other` have the same **day**.
-  bool eqvDay(DateTime other) => dateEqDay.eqv(this, other);
-
-  /// Return `true` when this [DateTime] and `other` have the same **year, month, and day**.
-  bool eqvYearMonthDay(DateTime other) => dateEqYearMonthDay.eqv(this, other);
-}
-
 /// Constructs a [DateTime] instance with current date and time in the local time zone.
 ///
 /// [IO] wrapper around dart `DateTime.now()`.
