+++ packages/fpdart/lib/src/task_option.dart
@@ -1,4 +1,5 @@
 import 'either.dart';
+import 'extension/option_extension.dart';
 import 'function.dart';
 import 'option.dart';
 import 'task.dart';
